requirements.txt
asgiref==3.8.1
Django==5.1.7
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
promise==2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
whitenoise==6.9.0
